doFS: use the apparent size instead of the disk usage
Using the disk size can lead to errors, as it seems like it takes some time 
for it to be the real file size (at least when using ZFS). As an example, 
here is the size of mfsroot.gz just after it's created and then after 1s:

28705	/root/mfsbsd/tmp/disk/mfsroot.gz
sleep 1;
31401	/root/mfsbsd/tmp/disk/mfsroot.gz

In order to solve this, use the apparent size instead of the disk size.