fix gpart usage in do_gpt.sh
Since FreeBSD commit r264378 partitions must be aligned to 4K, so 
specifying a start of 34 (17408 bytes) makes gpart yield and error an exit. 
Fix this by removing the -b argument at all, and just relying on gpart to 
place the freebsd-boot partition by itself in the first free space it finds.