This commit adds EFI boot support to mfsBSD.
EFI boot requires an EFI SYSTEM PARTITION with an EFI compable boot loader. The Makefile is modified to extract the FreeBSD EFI loader, create a ESP and add the ESP to both the ISO as well as the disk image.
Note that this allows an EFI system to boot into mfsBSD, but due to upstream limitations in FreeBSD, the resulting system will not work. See https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=251866.